requirements-ci.txt
# Lightweight CI requirements for GitHub Actions
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
websockets==12.0
redis==5.0.1
numpy==1.24.4